pyproject.toml
[tool.poetry]
name = "calendar"
version = "0.0.3"
description = "\"\""
authors = ["Kikoring <kyr4624@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry"}]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.36"
bcrypt = "^4.2.1"
datetime = "^5.5"
pyjwt = "^2.10.1"
api = "^0.0.7"
fastapi-mail = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
